DOC: Re-define convergence terms xtol and rtol for root solvers
Addresses #18740

Issue #18740 revealed behavior in bisect that was counterintuitive for
some, given the documentation of the xtol and rtol parameters; this is
equally applicable for most other root solver methods in _zeros_py.py.

The parameters xtol and rtol were previously defined in terms of
satisfying np.allsame; although that is correct, np.allsame is not
actually used in implementing these functions and one would need to
consult the documentation on np.allsame to see how to set xtol and rtol
in this other context. Now, these parameters are defined in terms of the
mathematical function that np.allsame applies; doing it this way is more
direct and about as concise.

Another approach would be to revise the tutorial; although the tutorial
on finding zeros of functions should be expanded, most of that expanded
language would not directly deal with convergence behavior.

This may close the issue.

[skip actions] [skip cirrus]