BLD: rename `setup.py` to `_setup.py` to signal it should no longer be used
See the added code comment in this commit for rationale of keeping `_setup.py`
around. We got another two bug reports that used `python setup.py install`
in the last day (gh-19022 and gh-19026), so it's time to make that
impossible.

For those who really really need it, they can still manually rename
`_setup.py` to `setup.py` for a while longer, but this should make
it much clearer that we no longer support this install method (beyond
the one conda-forge on Windows case).

[skip circle] [skip cirrus]