DEP: remove all deprecated `scipy.<numpy-func>` objects
As evidenced by gh-19065, changes to the contents of numpy's main
namespace or `__all__` dict have a direct impact on scipy's main
namespace. It may cause warnings, import errors in test suites
where warnings get converted to errors directly, and it changes
the actual contents of the `scipy` namespace.

The above was always the case, however in the run-up to NumPy 2.0,
we can expect a very large amount of changes in NumPy's main
namespace. We don't want to expose newly added NumPy functions
in the `scipy` namespace, nor have new breakage each time something
in numpy gets deprecated or removed. Hence, let's remove these
long-deprecated items now. Our deprecation message said "scipy 2.0"
as a way of saying "this is not urgent, these things will be removed
eventually". However, it has become urgent now. Changing this in
SciPy 1.12.0, which will roughly coincide with the NumPy 2.0 release
date, should be fine and healthy.