BUG: Fixes #19103 by adding back make_strictly_feasible to lsq trf.
PR #18896 removed make_strictly_feasible because it makes solutions
worse if they are close to the boundary.
However, the downstream algorithm that eventually solves the
optimization problem has some trouble with starting points that
are on the boundaries of a bounded optimization problem.
This either results in division by zero warnings, see #19102,
or in failure of finding the correct solution at all,
which is added as a regression test in this commit.
This PR brings back make_strictly_feasible but with the
smallest possible step from the boundary a floating point allows
by setting rstep=0 on the function make_strictly_feasible,
which uses np.nextafter to find such a point, hereby respecting
the strictly feasible point condition and minimizing boundary effects.
See strictly feasible points in interior-point methods for more context
https://nmayorov.wordpress.com/2015/06/19/trust-region-reflective-algorithm/