MAINT: signal: Remove the cval parameter from the private function _pad_test.
The parameter `cval` in the private helper function `_pad_test` is not
used, so remove it.  This change is inspired by a bug in Cython 3.0.1
(see SciPy issue gh-19132), and this change is a work-around for that bug,
but since the parameter is unused and in a private function, there is no
need to revert this change.