DOC: fix sparse docs "matrix" and "array" mixup (#18898)
* DOC: made sparse array/matrix docs only have array/matrix

Before, the docs for sparse arrays/matrices were written
for the array classes and "sparse array" was replaced with
"sparse matrix" to get the matrix class doc, but that did
not cover all strings that need to be replaced. Now, the
base docstrings have strings e.g. "{array|matrix}" that are
replaced with "array" or "matrix" to get the array or
matrix docs. Docstrings of the different sparse formats
are also made more consistent. See #18669.

* DOC: switch main sparse docs from "matrix" to "array"

Under "Usage information" in the main sparse doc page, the
matrix interface is still used; this commit switches it
all to the (recommended) array interface. Doc code
formatting is also improved. See #18669.

* DOC: change sparse class method docstrings to "array/matrix"

Prior to this, the method docstrings of sparse classes
(and their parent classes) were a mix of "array"s and
"matrix"s. This commit makes this consistent by changing
all relevant instances of "array" and "matrix" to
"array/matrix". When sparse matrices are deprecated,
change all instances of "array/matrix" to "array".
See #18669.

* DOC: move sparse class docs to array/matrix classes

Previously, the class docs of the sparse array/matrix
classes were defined in the base class, and then a string
replacement was used to obtain the array and matrix class
docs. To allow for more divergent array and matrix class
docs, it has been decided to simply define the docstrings
in the array/matrix classes, and this commit does that.
See #18669.

* DOC: change "csr_array/matrix" in code to "csr_array"

* DOC: remove "dense matrix" in sparse class docs

* DOC: improve BSR class docs

* DOC: "toarray()"" -> "toarray"

---------

Co-authored-by: Kat Mistberg <katmistberg@gmail.com>