MAINT: Fix/parametrize Rotation tests.
The test were slightly incorrect in the sens that they were testing that
at least one of the combination of angle/sequence/intrinsic-extrinsic
angle where remitting a Gimbal Lock warning, instead of each of them.

The warning is now checked _only_ and on _all_ on the line that are
supposed to emit such a warning.

And while I was at it I parametrized the tests, so each appear as a
separate inter when running with pytest, and extended to all the
parametrized test in case it ever have an issue somewhere.

In all of the `angles = np.array(...)` I have the impression that each
one should emit a warning and that angle should likely also be a
parameter, but it was starting to be a lot (maybe open an issue to
parametrize angle?)