ENH: Add ability to silence Gimbal Lock warnings for rotations.
While it is possible to silence warnings with `catch_warnings()` context
manager and a filter the approach has a few drawbacks:

1) The warnings are still emitted, which can have a perf impact
2) Using catch_warnings hits this bug `python/cpython#73858` that may
   reset a bunch of other warnings.

I also slightly refactor the code by moving the warning 1 level up into
`_compute_euler`
in the stack and having the core function not emit the warnings, but
return a tuple of (data, was_there_a_gimbal_lock). This in particular
also emit the warnings only once per function call instead of once per
gimbal lock.

Another possibility would be to move the warning one more level up into
`as_euler(...)` which is the only public method that uses
`_compute_euler(...)`. The other one being `_as_euler_from_matrix(...)`,
whcih is private and only use in tests. If I can to that, then I think
it might be ok to not add the `gimbal_lock_ok` keyword and just tell
users to reach for the private `_compute_euler`, and discard the warning
flag.

We could also return a boolean array of whcich rotations are Gimbal
Locked. I don't know if that would be useful.

Closes #19306