DOC: Rst don't like list without whitespaces.
It does not matter much as this is not rendered in the docs. But still
if something try to parse all the docstrings of all the function of
scipy it fails.

You can https://stsewd.dev/tree-sitter-rst/ to parse the docstring and
see how it fails.

And yes you want to not indent as otherwise it becomes a block quote.

[ci skip]