MAINT: remove support for building with setup.py
With Python 3.12 available and the Meson build mature, we don't need
this any longer. The last reason we kept this around is to allow
conda-forge to use it on Windows. That isn't possible anymore with
Python 3.12, and conda-forge builds on Windows with Meson are now close
enough to the finish line (a first build with Clang + Flang managed to
create a wheel, still work to do though).