fix pow method for sparrays with power 0
A**0 means every entry is 1, not just the originally non-zero entries.
So, the A**x function changes the nonzero structure to dense in a
discontinuous fashion near x=0. There is a question then as to whether
this method should return a sparse array with all entries 1, or return
a dense array, or raise an exception with a workaround described.

This PR implements raising an exception for `n` a zero valued scalar.
The message (and doc_string description of `n`) says if they want to
compute A**x for x==0, they should special case that value to use
`np.ones(A.shape, dtype=A.dtype)`