DOC: add `meson-python` as a build dependency under Building from source
The first line changed is a bug/omission, because the next step uses
`--no-build-isolation`. For the second line changed, `meson-python`
isn't actually required when using `dev.py`, however when an editable
install is tried, it is still needed. It's a minor change, so best
to always add it to build dependencies.

Closes gh-19565