BUG: interpolate/ndbspline: fix OOB for len(tx) != len(ty) in tensor product splines
To be able to release the GIL, we pack the tuple of 1D knot arrays
into a 2D array. Since the lengths of knot arrays for different
dimensions may differ, the 2D array has dimensions
(ndim, max_length), with extra elements filled with NaNs.

The actual lengths of knot arrays need to be known to the interval
search routine, which finds for a given `x`, the index `i` such that
`t[i] <= x < x[i+1]`.
Otherwise these extra nans confuse the interval search routine, leading
to OOB access down the line.