MAINT: interpolate/RGI: avoid unnecessary validation of data size (#19591)
RegularGridInterpolator allows changing the interpolation method
at call site; if it is changed, data sizes need re-validation for
a new method (cubics require at least four points and so on).

At the call site, `__call__(..., method=None)` means that the interpolation
method is the same as self.method and thus there is no need to re-validate
self.values.

RGI's interpolaton method (linear, cubic, etc) can be supplied in both
`__init__` and `__call__`. The former is stored on the instance, and
the latter allows changing the method for this particular call.
Further, `__call__(..., method=None)` means 'use self.method'.