CI: undo cython==0.29.35 pin for the 32-bit Linux job, remove `!=3.0.3` [wheel build]
We're at Cython 3.0.6 now and the relevant upstream bugs have been fixed.
We no longer need the `!=3.0.3` condition in CI jobs, because that version
will never be installed anymore. We only keep it for now in `pyproject.toml`
and in `environment.yml`, because there it is still useful information
for downstream redistributors.