TST: fix incorrect signal.sosfilt tests
This caused failures on macOS arm64 with NumPy main, because
clongdouble arrays started to be passed to `np.linalg.eigvals`.

This code, as the deleted comment in `cast_sos2tf` said, does not
actually work with long double dtypes, nor with object. So trying
to cast to complex128, call the functionality, then cast back,
doesn't really make sense as a test. Hence simply delete the unsupported
dtypes from the tests.