TST: add `norecursedirs` to pytest.ini, avoid picking up vendored tests
The ones in the highs git submodule were actually causing `pytest scipy`
to error on discovery. The other added dirs don't contain relevant
tests, and it may help to avoid searching their trees (numpy does this
too for `doc` and `tools`).

Note that the `benchmarks` directory actually does contain some tests,
and they pass when run with a plain `pytest` in the root of the repo.