BLD: require Cython >=3.0.4, drop 0.29.X support
This will make maintenance a bit easier, and it's time; NumPy
dropped Cython 0.29.X support months ago and this has gone fine.

This also fixes an annoying build warning that shows up twice, namely:

```
[5/7] Generating 'scipy/_lib/_ccallback_c.cpython-310-x86_64-linux-gnu.so.p/_ccallback_c.c'
performance hint: /home/rgommers/code/scipy/scipy/_lib/_ccallback_c.pyx:138:82: Exception check will always require the GIL to be acquired. Declare the function as 'noexcept' if you control the definition and you're sure you don't want the function to raise exceptions.
```

The code is correct and not easy to change. It's only test code as well.
The directive to silence the unhelpful hint was introduced in Cython
3.0.4, hence that's used as the new minimum version.