ENH: single-pass CSR sparse matrix binops
The CSR elementwise binary operators have two implementations: one for canonical and another for general formats. This merges the format check with the canonical format implementation. Now optimistically execute the canonical implementation with a fallback to the general if the operands are not canonical. This eliminates a full memory scan of each operand, which speeds up this memory-bound operation.