WIP: ENH: enable ILP64 BLAS/LAPACK support again
This was left out initially when adding support for building with
Meson. The numpy.distutils based build supported ILP64 builds with
OpenBLAS and its `64_` symbol suffix scheme. We now add this back
and make it more generic, so that it also supports the `_64` scheme
of MKL as well as the `$NEWLAPACK_ILP64` scheme from Accelerate.