TST: move reference data for test_real_transforms to a fixture
Move the code for getting the reference data in test_real_transforms
to a fixture.  This easily ensures that the data is loaded only once
(the previous caching logic did not work correctly), and it makes it
possible to easily close the data files afterwards.

This also fixes test flakiness on PyPy.  Unlike CPython, PyPy does not
aggressive GC open files.  Therefore, when every test iteration opened
the files again, the number of open files quickly grew to the system
limit and caused further `open()` calls to fail.

Fixes #19553