MAINT: make isinstance check in `stats._distn_infrastructure` more robust
This check for a list return type from `atleast_1d` may fail if
in NumPy 2.0 the return type will be tuple rather than list. Also,
for PyTorch it's tuple, and in general tuple seems preferred here
(due to immutability).

Checking for both list and tuple makes the code future-proof, and
only costs a few tens of nanoseconds more (performance-wise this line
doesn't matter).

[skip cirrus] [skip circle]