ENH: Set up specific logger for scipy.optimize.shgo.
Set the modules directly associated with SHGO to log to the specific
logger named "scipy.optimize.shgo" instead of the "root" logger. This
allows the user to specify the logging level
(and format, output files etc.) for only this routine without affecting
loggers that may be set up in other parts of the user's code. When
logging to the "root" logger, any changes made to the logger will affect
all other loggers as they all inherit from "root".