TST: stats: fix a few issues with non-reproducible seeding
The use of `hash('a_descriptive_string')` is cute, but unfortunately
hashes are not reproducible at all. This resulted in a hard to debug
failure in gh-19797. So remove all usages of `hash()`, and also
use `np.random.RandomState` for good measure since that has more
stability guarantees than `np.random.default_rng`.