Fix AIX build break.
This patch is a fix to AIX build break that happened due to a variable defined as hz in _amos.c.

In AIX, stdlib.h imports m_params.h that defines a macro hz.
This results in the below error in AIX which is why hz variable is renamed to half_z.

FAILED: scipy/special/lib_amos.a.p/_amos.c.o
In file included from /usr/include/sys/context.h:33,
                 from /opt/freeware/lib/gcc/powerpc-ibm-aix7.3.0.0/10/include-fixed/sys/signal.h:374,
                 from /opt/freeware/lib/gcc/powerpc-ibm-aix7.3.0.0/10/include-fixed/sys/wait.h:59,
                 from /opt/freeware/lib/gcc/powerpc-ibm-aix7.3.0.0/10/include-fixed/stdlib.h:396,
                 from ../scipy/special/_amos.c:95:
../scipy/special/_amos.c: In function 'amos_seri':
../scipy/special/_amos.c:4203:45: error: expected identifier or '(' before numeric constant
 4203 |     double complex ak1, ck, coef, crsc, cz, hz, rz, s1, s2, w[2];