MAINT: Use lazy logging in scipy.optimize.shgo
Modify logging statements currently using f-strings to use lazy logging
instead. This avoids producing the log statement unless it is actually
going to be logged, giving a significant increase in performance when
the messages are not logged. This eliminates the performance-related
incentive to hide log statements behind conditionals or leave them
commented out.