BLD: fix issue with escape sequences in `__config__.py`
Reported in https://github.com/conda-forge/scipy-feedstock/issues/264:
linker args contained a `\s` on Windows as part of a path in a linker
flag.

This is basically the same bug as gh-17936, it was fixed for paths
in `__config__.py` but not for things like linker flags which usually
don't contain a path but may do so.

Now turn everything except name/version/found/id, which really should
not contain escapes, into raw strings.

[skip circle] [skip cirrus]