DOC: update pydata-sphinx theme (#16660)
Pick up improvements from several recent pydata-sphinx-theme (PST) releases,
for example around search bars, accessibility, UI improvements (e.g. correct
visibility of components), while staying close to the existing styling of the
docs (aside from minor updates of switcher button, logo placement, etc.).

This was held up for an extended period due to build time regressions, as well
as other issues that arose. The result of various investigations (and much help
from upstream devs! 🙏) is that we now require PST>=0.15.2.

However, it was discovered that the even the most recent PST version is not
thread-safe currently, so we need to build single-threaded, in order to ensure
that (i.a.) the TOC sidebar is shown correctly on all pages.

Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>
Co-authored-by: Daniel McCloy <dan@mccloy.info>
Co-authored-by: Rambaud Pierrick <12rambau@users.noreply.github.com>