ENH: stats.gamma.fit: use analytical formulas (#19932)
* ENH: use explicit formula for scipy.stats.gamma.fit('mm')
* TST: add test case for scipy.stats.gamma.fit('mm')
* TST: add "round-trip" test for gamma.fit('mm').

The test generates a random sample from a gamma distribution and
fits a gamma distribution to it.  It then checks that the first r
moments of the fitted distribution is equal to those of the sample,
where r is the number of estimated parameters.

The simulation parameters are restricted to a small-ish range to
ensure the MM estimates lie within the range of valid parameter
values, so that "round-trip" testing can be performed.