TST, MAINT: test_immediate_updating fix
* Fixes #19848

* Safely forking from a multithreaded process is problematic
and deprecated in Python 3.12+; the test failure in the linked
issue is fully reproducible on x86_64 Linux when using `pytest-xdist`
and more than 1 process with Python `3.12`.

* Andrew's solution seemed the most effective in my hands; I did also
explore using one of the environment variables that tells you if you're
in a `pytest-xdist` worker, but the code seemed messier without
much benefit.

* The alternative start method used here (`spawn`) is slower of course,
as the docs indicate. For 1 core, the test runs in 0.02 s locally, and
the worst I saw was 0.12 s with 32 cores (using the timing from
`--durations`, not overall suite time of course) vs. 0.07 s with
32 cores using `fork` (both methods are 0.02 s locally with 2
`xdist` cores, I suppose that's closest to what we'd do in CI)

* the `solver` variable name was shortened to accommodate line length
(I was `-0.5` on that but linter was complaining)

[skip cirrus] [skip circle]