BUG: loglaplace moment should be non-negative.
The current implementation of scipy.stats.loglaplace.moments incorrectly
returns a negative number when the requested moment does not exist. This
commit fixes this bug and correctly returns np.inf instead.

A test case is added for parameter combinations that lead to non-finite
moments. Non-central moment is allowed to return inf or nan as it is not
well-defined when a lower-order non-central moment is non-finite.

The generic stats() implementation is updated to suppress floating point
warnings caused by (inf-inf) operation.