TST: add sanity check to _assert_less_or_close_loglike().
The sanity check asserts that the analytical fit result is not identical
to the numerical fit result.  This is usually true due to floating point
arithmetic.

The check may be bypassed by passing maybe_identical=True.  This could
happen if the analytical routine falls back to numerical MLE in some
cases, or if the objective function is simple enough that the numerical
solution and the analytical solution coincide.