MAINT: stats: fix test failure in `kendalltau_seasonal`
This failure shows up when `SCIPY_ARRAY_API` is set:
```
FAILED scipy/stats/tests/test_mstats_basic.py::TestCorr::test_kendalltau_seasonal - TypeError: Inputs of type `numpy.ma.MaskedArray` are not supported.
```

What happens in the code is that a masked array was passed to
`special.erfc`. That function (and `special` as a whole) is not aware
of masked arrays and hence silently drops the mask. That was fine in
this case because the masked array being passed always had a mask with
only False values. So extracting the underlying data and passing only
that changes nothing else, while avoiding the failure and implicit
behavior.

[skip cirrus] [skip circle]