TST: optimize: silence the output from calling cobyla with disp=True
Not doing this prints pretty verbose output, which pytest then displays
below the test summary at the end of a run of the test suite.

`capfd` and `capsys` are built-in pytest fixtures to capture such
output. Note that they are not robust, and you have to choose between
the "file descriptor" and "sys.stdout/stderr" modes, so when we have
code that prints from both Python and Fortran, it doesn't seem possible
to silence everything.

Note also that some output shows up with `python dev.py -s optimize`
but not with `python dev.py -t scipy.optimize.tests`, so it's hard to
track down.

The one removed test was already xfail-ed, still noisy, and not useful.

I also considered removing or changing some of the printing code, but
decided against that because that code is likely to go away with the
"remove Fortran 77 code" efforts.

[skip cirrus] [skip circle]