BUG: smoke-docs: fix dev.py test --doctest -t path/to/file
plumb through the doctest collection `strategy`:

- dev.py test --doctests : uses strategy=api
- dev.py -s submodule : uses strategy=api
- dev.py -t scipy/linalg._basic.py : uses strategy=None

This sorta makes sense: if you want to smoke-test a package or a subpackage, only
look at the public API --- this is not changed.

If you smoke-test an _individual file_, you want functions in that file, so the
vanilla doctest finder gives you all of them. IOW, given a single implementation
file, you have no way to tell which functions from this file end up in the scipy
public API.

A possible UX glitch is that `dev.py -t` may pick up extra docstrings (from things
which remain private in SciPy). It may thus report some extra failures,
compared to `dev.py -s submodule/which/has/the/file` --- if the docstrings of these
private objects are not perfect.

Two options: bend over backwards to invent a way to figure out if an object from
a private _module ends up in the public API, or simply document it and move on.
I suspect the former will be hacky and unreliable, so am opting for the latter.

Yet another way around it is to specify an individual object:

```
$ python dev.py test --doctests -t scipy/linalg/_basic.py::scipy.linalg._basic.det
```

note the need to specify the full dotted path.