BLD: remove use of `NPY_VISIBILITY_HIDDEN`
This usage in `sparsetools` predates the move to Meson; Meson
hides symbols by default (see gh-15996). I've checked that there
is no change in binary size on Linux with GCC and macOS with Clang.

Use of `NPY_VISIBILITY_HIDDEN` is not only a minor cleanup, but
technically more correct - in case SciPy gets built with a different
compiler than NumPy, the value of this define may not be correct
(most but not all compilers will support the non-standardized
`__attribute__((visibility("hidden")))`.

We shouldn't need visibility attributes, but in case we want to
reintroduce them in the future, it's better to copy the build-time
compiler check for support of the attribute, rather than reuse the
result of the NumPy check.

[skip circle]