BUG: optimize: `NewtonCG` `min` crashes with `xtol=0` (#20299)
* BUG: Optimize: NewtonCG min crashes with xtol=0

The minimize function crashes if `xtol` is set to `0`. This is
because the initial value of a loop variable is `2 * xtol`, which
is intended to ensure the loop runs at least once. However, if
`xtol = 0` then the loop never runs, a variable isn't instantiated
that is used outside of the loop, and an `UnboundLocalError` is
raised.

To fix this, the initial value of the loop variable is set to the
max float value.

* BUG: Optimize: UnboundLocalError for xtol=0 NCG

Corrected error from previous rework when switching to
`np.linalg.norm` that introduced the possibility of the unbound
variable error.

* TST: Optimize: NewtonCF Minimize when xtol=0

Regression test for gh-20214 which addressed an issue in the
Newton CG method which would raise an UnboundLocalVariable error
if `xtol` was set to zero. It also ensures the method finds the
solution in this case.

* MAINT: Add comment to explain initial value

Comment added to explain the purpose of setting `update_l1norm` to the max float value - this is to ensure the preceding while loop executes at least once.

Co-authored-by: Christian Lorentzen <lorentzen.ch@gmail.com>

* TST: Test solution value as well as success

This is for the `xtol=0` test in the Newton CG minimize method.

* TST: Explicitly extract val from array

---------

Co-authored-by: Christian Lorentzen <lorentzen.ch@gmail.com>