BUG: Fix for scipy.special seterr, geterr, errstate (#20321)
This is an alternative to gh-20316. `scipy.special.seterr`, `scipy.special.geterr`,
`scipy.special.errstate` currently do not work for ufuncs not implemented
in the extension module `_ufuncs` (whose Cython code is generated by
`_generate_pyx.py`). This is because `sf_error.c` manages an error state in a
static array, and currently each extension module has a separate copy of
`sf_error.c` compiled in, when they actually all should be sharing this state.
As identified by @izaid, a potential solution is to manage this state in a
shared library. This PR differs from gh-20316 in that the shared library
only manages the shared state and nothing else. I think it's best to go
with the minimal set of changes for now in order to leave more options
open for the future.

Shared libraries are tricky for packaging. This is the first shared library
that is part of SciPy - a regular shared library that is, not a Python
extension module. On most platforms things work out of the box
with the added `$ORIGIN` RPATH, however Windows does not have
RPATH support and hence pre-loading the DLL before importing any
extension modules that depend on it is necessary.
This also requires some tweaks to `delvewheel` usage. It's expected
that this works transparently for contributors and redistributors
on Windows.

Co-authored-by: Ralf Gommers <ralf.gommers@gmail.com>