BUG: fix jensenshannon returns NaN for some similar distributions
Closes #20083

The jensenshannon function returns NaN values for some distributions
close to each other by taking the square root of a negative number.
The negative numbers appear in the return of the rel_entr function.
To fix this, use the equivalent kl_div function instead of rel_entr
because it always returns non-negative values. This is possible
because the extra parameters on the kl_div formula that make the
values non-negative will be canceled out in the sum.