DOC: stats: speed up the very slow `bootstrap` examples
This takes a few percent of the total doc build time (which
is painfully long). Using 9999 resamples several times is excessive,
so this trims things down a bit. The result is still taking
longer to execute than desirable (~5 sec on my machine),
but this is already 5x better so enough for now.

[docs only]