BLD: avoid setting an environment variable in a meson.build file
This was introduced last month in gh-20232; I commented on it there
about it maybe being able to cause an issue later. And it just did,
see gh-20433. Upgrading Meson in an existing build environment is
technically unsupported by Meson, but has always "just worked". Now
we hit a corner case on Windows, because when we set an environment
variable in a `meson.build` file, a pickled version of Meson is called
and this is less robust to upgrading Meson in the build env.

The fix is a bit more verbose, but should be reliable - the `importlib`
importing is directly copied from how it's done in NumPy's code
generation scripts.