TST: interpolate: bump test tolerance in test_rbfinterp::test_pickleable
An exact equality was reported as flaky on conda-forge in
https://github.com/scipy/scipy/issues/20472

The tolerance violations are of the order of fp noise (< 2e-16), and I don't
think that pickling/unpickling guarantees bit-to-bit compatibility.
In principle, this may invoke recalculations and those may be subject
to fp noise. So I think it's OK to only require allclose(atol=eps)
instead of exact equality.