BLD: default to C17 rather than C99
This update should be a safe bump, because C17 is a pure maintenance
update of C11 with no new features; and C11 is mostly a
maintenance/bugfix update of C99 with a few new features.
Our [toolchain roadmap](http://scipy.github.io/devdocs/dev/toolchain.html#c-compilers),
which was updated just days ago, says that we can use C17
given our minimum compiler versions.

I also note that in numpy#25072, the numpy default was updated from
C99 to C11 and that went fine (the one change needed was a Cython bug
fix, which went into Cython 3.0.6 - and our minimum version is currently
3.0.8).

Finally, another motivation to do this now is that under the
`--disable-gil` build of CPython 3.13, Cython currently generates code
that uses `static_assert`, which is a C11 feature. Hence our current
default of `-std=c99` doesn't build with that nogil build.