DOC: Use vendored MathJax distribution as fallback
This commit adds a small code snippet to check
for an internet connection by pinging Google's DNS.
If an internet connection is available, it will use
the vanilla MathJax distribution, and if there isn't,
then the local (vendored) MathJax distribution will
be used. This ensures that the local one is used as
just a fallback with a mechanism to check for the
availability of the usual distribution.