MAINT: vendor Tempita in `scipy/_build_utils`
This is an unmodified copy from `Cython/Tempita/` as of commit
`023d4af35` in Cython's master branch (21 April 2024).
Tempita hasn't been maintained independently on PyPI for 10+ years;
we've used the Cython version which is semi-public (undocumented)
for a while, with the note that we should vendor it if that ever
fails. It now failed once (see scipy#20535), and conceptually that
makes sense - this is the only place where we don't invoke the
`cython` executable but actually do `import Cython`. That can fail
in a distro setup where Cython is installed for a different Python
interpreter than the active one.

The license file was taken over from the `maintenance/1.26.x` branch
in the `numpy` repo - which came from the original upstream repo
for Tempita (now disappeared) as discussed in numpy#8096.
See scipy#20572 for more clarification about the MIT license that
this code is under.