ENH: Add more initialization methods for HessianUpdateStrategy (#13534)
* Remove init_scale restriction to scalar

* Remove init_scale restriction to scalar

* ENH: add arbitrary shaped initial scale for HessianUpdateStrategy

Remove the float conversion that implies a scalar. This adds a lot of flexibility to e.g. intialize with an array or even a matrix. If a matrix is given, it will be used as the initial matrix.

* ENH: add arbitrary shaped initial scale for HessianUpdateStrategy

Remove the float conversion that implies a scalar. This adds a lot of flexibility to e.g. intialize with an array or even a matrix. If a matrix is given, it will be used as the initial matrix.

* ENH: add arbitrary shaped initial scale for HessianUpdateStrategy

Remove the float conversion that implies a scalar. This adds a lot of flexibility to e.g. intialize with an array or even a matrix. If a matrix is given, it will be used as the initial matrix.

* ENH: add arbitrary shaped initial scale for HessianUpdateStrategy

Remove the float conversion that implies a scalar. This adds a lot of flexibility to e.g. intialize with an array or even a matrix. If a matrix is given, it will be used as the initial matrix.

* ENH: add arbitrary shaped initial scale for HessianUpdateStrategy

Remove the float conversion that implies a scalar. This adds a lot of flexibility to e.g. intialize with an array or even a matrix. If a matrix is given, it will be used as the initial matrix.

* ENH: add arbitrary shaped initial scale for HessianUpdateStrategy

Remove the float conversion that implies a scalar. This adds a lot of flexibility to e.g. intialize with an array or even a matrix. If a matrix is given, it will be used as the initial matrix.

* ENH: add arbitrary shaped initial scale for HessianUpdateStrategy

Remove the float conversion that implies a scalar. This adds a lot of flexibility to e.g. intialize with an array or even a matrix. If a matrix is given, it will be used as the initial matrix.

* docs: improve HessianUpdateStrategy parameter init_scale

* ENH: make init_scale in HessianUpdateStrategy a real array

This will fail loudly if it can't be cast to the exact dtype, e.g. if it is complex. Added an extra check to ensure that this will do. Also extended the tests to check for both approximation types, hess and inv_hess

* ENH: add legacy cast in HessianUpdateStrategys init_scale

Added the float cast again to ensure full legacy behavior

* FIX: Check if init_scale is str in HessianUpdateStrategy

Gives FutureWarning otherwise as a string comparison with an array

* TEST: Check if init_scale is str in HessianUpdateStrategy test

Gives FutureWarning otherwise as a string comparison with an array

* REFACTOR: Make elif for if that would have been elif anyway

* chore: remove whitespaces

* chore: remove whitespaces

* chore: remove unused import

* enh: add input checks

* tests: add match check in message

* tests: use pytests raises

* tests: typo

* fix: init multiply

* fix: init multiply

* style: fix

* tests: escape regex

* tests: escape regex 2

* tests: escape regex 2

* tests: escape regex 3

* tests: escape regex 4

* tests: escape regex 4

* tests: fix test with correct error

* enh: add more descriptive errors

* enh: add more descriptive errors

* enh: be explicit about copy

* chore: reformat

* tests: fix exceptions

* chore: fix linting

* chore: fix linting 2